pyproject.toml
[tool.poetry]
name = "alpaca-trading"
version = "0.1.0"
description = ""
authors = ["roberte777 <robewilkes@yahoo.com>"]
readme = "README.md"
packages = [{include = "alpaca_trading"}]

[tool.poetry.dependencies]
python = "^3.11"
alpaca-py = "^0.14.0"
numpy = "^1.26.3"
pandas = "^2.2.0"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
